// src/model/JointType.ts
export enum JointType {
  Neck = 0,
  Nose = 1,
  CenterHip = 2,
  LeftShoulder = 3,
  LeftElbow = 4,
  LeftWrist = 5,
  LeftHip = 6,
  LeftKnee = 7,
  LeftAnkle = 8,
  RightShoulder = 9,
  RightElbow = 10,
  RightWrist = 11,
  RightHip = 12,
  RightKnee = 13,
  RightAnkle = 14,
  RightEye = 15,
  LeftEye = 16,
  RightEar = 17,
  LeftEar = 18
}